Pick reasonable values for axis ends
Thanks to Drew Noakes and his algorithm posted at
http://stackoverflow.com/questions/326679/choosing-an-attractive-linear-scale-for-a-graphs-y-axis
which I adapted into g.snapEnds(). In my case, with a time-series
dataset, the values for the axis were being selected outside the
timeframe of the data. Therefore, the graph was plotted outside the
graph area on both sides of the graph. This algorithm selects axis
ends that are outside the dataset, is simple, and likely much faster