enabling hrefs if the pie chart has a single slice
Adding code to the branch handling single sliced pie chart to enable
click through. It looks like a single slice has a cover as a separate
object, but multiple slices don't have separate cover, not sure why that
is. The cover for the single slice stays on top so it will get the href
attr.