giving access to cut and defcut with properties maxSlices and minPercent
Removing the hardcoded limitation of cut = 9 and exposing it as
maxSlices. This sets the max for individual slices, all other values
will be collapsed.
Deriving defcut flag from the minPercent option. If it's set to 0 no
attempts to prevent rendering very small slices will be made. Otherwise,
piechart will collapse all slices below this threshold. The default is
1.