Update drawing of pie chart when there are single or no values
* When the length of data is 1, it draws a path instead of a circle so
hover functionality to use this.sector will work the same as if there
are multiple slices.  (line 85 - 102)
* Fix drawing when a single value equals the total.  Prior, it did not
draw the chart.  This change draws a circle path. (line 134, 147-162)
* Added an additional option to ignore the min percentage and show all
values.  Passing in a min-percent as 0 to show all values and
evaluating Boolean( minPercent ) is still false, so it didn’t show all
values.  Passing in in opts.ignoreMinPercent allows you to evaluate it
correctly. (line 61)